Tests/LibWeb/Text/input/selection-on-rendered-text.html

<!DOCTYPE html>
<script src="include.js"></script>
<p id="text">Lorem     ipsum dolor sit amet</p>
<script>
    test(() => {
        internals.doubleclick(85, 20);

        const selection = document.getSelection().getRangeAt(0);
        println(`Selected range: ${selection.startOffset} - ${selection.endOffset}`);
    });
</script>
