Tests/LibWeb/Text/expected/selection-on-rendered-text.txt

Selected range: 6 - 11
